[sp] update & refactor distribution constants
