+++ mage_ai/frontend/components/datasets/overview/constants.ts
@@ -1,5 +1,3 @@
-import { ColumnTypeEnum } from '@interfaces/FeatureType';
-
 export const TABS_QUERY_PARAM = 'tabs[]';
 export const SHOW_COLUMNS_QUERY_PARAM = 'show_columns';
 export const COLUMN_QUERY_PARAM = 'column';
@@ -12,9 +10,3 @@ export const TABS_IN_ORDER = [
   TAB_VISUALIZATIONS,
   TAB_DATA,
 ];
-
-export const COLUMN_DISTRIBUTION_STATS = {
-  [ColumnTypeEnum.EMAIL]: 'domain_distribution',
-  [ColumnTypeEnum.TEXT]: 'word_distribution',
-  default: 'value_counts',
-};
